Move method enum out of client class
